# notes/stylenest-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/cart/items/{cartItemId}": {
            "put": {
                "tags": [
                    "cart-controller"
                ],
                "operationId": "updateQuantity",
                "parameters": [
                    {
                        "name": "cartItemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCartRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseCartResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "cart-controller"
                ],
                "operationId": "removeItem",
                "parameters": [
                    {
                        "name": "cartItemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/address/{id}": {
            "get": {
                "tags": [
                    "address-controller"
                ],
                "operationId": "getAddressById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAddressResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "address-controller"
                ],
                "operationId": "updateAddress",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddressRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAddressResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "address-controller"
                ],
                "operationId": "deleteAddress",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/address/{id}/default": {
            "put": {
                "tags": [
                    "address-controller"
                ],
                "operationId": "setDefaultAddress",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/cart/add": {
            "post": {
                "tags": [
                    "cart-controller"
                ],
                "operationId": "addToCart",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddToCartRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseCartResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/verify-otp": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "verifyOtp",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyOtpRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseString"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/reset-password": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "resetPassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseString"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/forgot-password": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "forgotPassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseString"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/address": {
            "get": {
                "tags": [
                    "address-controller"
                ],
                "operationId": "getAddresses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListAddressResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "address-controller"
                ],
                "operationId": "addAddress",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddressRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAddressResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/wishlist/items/{wishlistItemId}": {
            "delete": {
                "tags": [
                    "wishlist-controller"
                ],
                "operationId": "removeItem_1",
                "parameters": [
                    {
                        "name": "wishlistItemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {}
                }
            },
...
            "UpdateCartRequest": {
                "type": "object",
                "properties": {
                    "quantity": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    }
                },
                "required": [
                    "quantity"
                ]
            },
            "ApiResponseCartResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/CartResponse"
                    }
                }
            },
...
            "CartResponse": {
                "type": "object",
                "properties": {
                    "cartId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CartItemResponse"
                        }
                    },
                    "totalPrice": {
                        "type": "number"
                    },
                    "totalItems": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "AddressRequest": {
                "type": "object",
                "properties": {
                    "fullName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "phone": {
                        "type": "string",
                        "minLength": 1
                    },
                    "phoneCountryCode": {
                        "type": "string"
                    },
                    "addressLine1": {
                        "type": "string",
                        "minLength": 1
                    },
                    "addressLine2": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string",
                        "minLength": 1
                    },
                    "state": {
                        "type": "string",
                        "minLength": 1
                    },
                    "country": {
                        "type": "string",
                        "minLength": 1
                    },
                    "countryCode": {
                        "type": "string"
                    },
                    "postalCode": {
                        "type": "string",
                        "minLength": 1
                    },
                    "addressType": {
                        "type": "string",
                        "enum": [
                            "HOME",
                            "OFFICE",
                            "OTHER"
                        ]
                    },
                    "isDefault": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "addressLine1",
                    "addressType",
                    "city",
                    "country",
                    "fullName",
                    "phone",
                    "postalCode",
                    "state"
                ]
            },
            "AddressResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "phoneCountryCode": {
                        "type": "string"
                    },
                    "addressLine1": {
                        "type": "string"
                    },
                    "addressLine2": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "countryCode": {
                        "type": "string"
                    },
                    "postalCode": {
                        "type": "string"
                    },
                    "addressType": {
                        "type": "string",
                        "enum": [
                            "HOME",
                            "OFFICE",
                            "OTHER"
                        ]
                    },
                    "isDefault": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ApiResponseAddressResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AddressResponse"
                    }
                }
            },
...
            "AddToCartRequest": {
                "type": "object",
                "properties": {
                    "productVariantId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "quantity": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    }
                },
                "required": [
                    "productVariantId",
                    "quantity"
                ]
            },
            "VerifyOtpRequestDTO": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    },
                    "otp": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "email",
                    "otp"
                ]
            },
            "ApiResponseString": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "string"
                    }
                }
            },
            "ResetPasswordRequestDTO": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    },
                    "newPassword": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[^a-zA-Z0-9]).{8,64}$"
                    }
                },
                "required": [
                    "email",
                    "newPassword"
                ]
            },
            "RegisterRequest": {
                "type": "object",
                "properties": {
                    "fullName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[^a-zA-Z0-9]).{8,64}$"
                    },
                    "phone": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "fullName",
                    "password"
                ]
            },
            "AuthResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    }
                }
            },
            "LoginRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "ForgotPasswordRequestDTO": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    }
                },
                "required": [
                    "email"
                ]
            },
...
            "ApiResponseListAddressResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AddressResponse"
                        }
                    }
                }
            }
        }
    }
}